requirements.txt
psycopg2-binary==2.9.3
requests==2.28.1
vk-api==11.9.9

selenium~=4.4.3